Changed defines to enums in rfm70

+++ lib430/src/libair/rfm70.c
@@ -1,5 +1,5 @@
 #include "rfm70.h"
 
-int rfm70_initialize( RFM70 *radio,	GPIOPIN pClock, GPIOPIN pDataInBi, GPIOPIN pDataOut, long options) {	
+int rfm70_initialize( RFM70 *radio,	GPIOPIN pClock, GPIOPIN pDataInBi, GPIOPIN pDataOut, int options) {	
 	return 0;
 }
